emojigen: Refactor into modules
Add subcommands to emojigen and split into modules to reuse common
functionality. The 'source' subcommand replaces the previous
functionality, the 'groupnames' and 'translate' subcommands are stubs
for now.
