+++ tools/emojigen/src/groupnames.c
@@ -1,0 +1,11 @@
+#include "groupnames.h"
+
+#include <stdlib.h>
+
+int dogroupnames(int argc, char **argv)
+{
+    (void)argc;
+    (void)argv;
+    return EXIT_FAILURE;
+}
+
