+++ tools/emojigen/src/groupnames.h
@@ -1,0 +1,6 @@
+#ifndef EMOJIGEN_GROUPNAMES_H
+#define EMOJIGEN_GROUPNAMES_H
+
+int dogroupnames(int argc, char **argv);
+
+#endif
