+++ tools/emojigen/src/sourcegen.h
@@ -1,0 +1,6 @@
+#ifndef EMOJIGEN_SOURCEGEN_H
+#define EMOJIGEN_SOURCEGEN_H
+
+int dosource(int argc, char **argv);
+
+#endif
