+++ tools/emojigen/src/translate.c
@@ -1,0 +1,11 @@
+#include "translate.h"
+
+#include <stdlib.h>
+
+int dotranslate(int argc, char **argv)
+{
+    (void)argc;
+    (void)argv;
+    return EXIT_FAILURE;
+}
+
