+++ tools/emojigen/src/translate.h
@@ -1,0 +1,6 @@
+#ifndef EMOJIGEN_TRANSLATE_H
+#define EMOJIGEN_TRANSLATE_H
+
+int dotranslate(int argc, char **argv);
+
+#endif
